pyproject.toml
[tool.poetry]
name = "social-awareness-online"
version = "0.1.0"
description = "In this project, we will harness the power of Python to perform sentiment analysis on social media posts related to a specific social justice issue, such as #BlackLivesMatter or #ClimateJustice. The primary objective is to analyze public sentiment and visualize the data to gain insights into how people perceive and discuss these critical topics."
authors = ["DeepeshKalura <deepeshkalurs@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tweepy = "^4.14.0"
textblob = "^0.18.0.post0"
pandas = "^2.2.2"
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
python-dotenv = "^1.0.1"
kaggle = "^1.6.12"
jupyterlab = "^4.2.0"
streamlit = "^1.34.0"
openai = "^1.30.1"
langchain = "^0.2.0"
geolocator = "^0.1.1"
geocoder = "^1.38.1"
langchain-openai = "^0.1.7"
duckduckgo-search = "^6.1.0"
langchain-community = "^0.2.0"
langchainhub = "^0.1.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
